Sounds Added

+++ FatBall/Assets/Scripts/MonstersSpawnerControl.cs
@@ -12,7 +12,7 @@ public class MonstersSpawnerControl : MonoBehaviour {
 
     public GameObject monster;
     public int num_of_monsters = 0;
-    public int monsters_limit = 3;
+    public float monsters_limit = 3;
 
     // Use this for initialization
     void Start () {
